Fixed tests and rename
